split stack trace implementations per implementation; code cleanup; unittest
