+++ src/daemon/buildinfo.c
@@ -1303,7 +1303,7 @@ __attribute__((constructor)) void initialize_build_info(void) {
 #ifdef HAVE_LIBMNL
     build_info_set_status(BIB_LIB_LIBMNL, true);
 #endif
-    build_info_set_value(BIB_LIB_STACKTRACE, capture_stack_trace_backend());
+    build_info_set_value(BIB_LIB_STACKTRACE, stacktrace_backend());
 
 #ifdef ENABLE_PLUGIN_APPS
     build_info_set_status(BIB_PLUGIN_APPS, true);
