+++ src/daemon/daemon.c
@@ -428,7 +428,7 @@ int become_daemon(int dont_fork, const char *user) {
         // the child
         gettid_uncached();
         nd_initialize_signals(false);
-        capture_stack_trace_flush();
+        stacktrace_flush();
 
         // become session leader
         if (setsid() < 0) {
@@ -451,7 +451,7 @@ int become_daemon(int dont_fork, const char *user) {
         // the child
         gettid_uncached();
         nd_initialize_signals(false);
-        capture_stack_trace_flush();
+        stacktrace_flush();
     }
 
     // generate our pid file
