+++ src/libnetdata/log/nd_log-init.c
@@ -136,7 +136,7 @@ void nd_log_initialize_for_external_plugins(const char *name) {
 
     //    nd_log(NDLS_COLLECTORS, NDLP_NOTICE, "FINAL_LOG_METHOD: %s", nd_log_id2method(method));
 
-    capture_stack_trace_init();
+    stacktrace_init();
 }
 
 // --------------------------------------------------------------------------------------------------------------------
@@ -266,7 +266,7 @@ void nd_log_initialize(void) {
     for(size_t i = 0 ; i < _NDLS_MAX ; i++)
         nd_log_open(&nd_log.sources[i], i);
 
-    capture_stack_trace_init();
+    stacktrace_init();
 }
 
 void nd_log_reopen_log_files(bool log) {
@@ -284,12 +284,12 @@ int nd_log_systemd_journal_fd(void) {
 }
 
 void nd_log_reopen_log_files_for_spawn_server(const char *name) {
-    nd_log_forked = true;
     nd_log.fatal_hook_cb = NULL;
     nd_log.fatal_final_cb = NULL;
 
     gettid_uncached();
-    capture_stack_trace_flush();
+    stacktrace_flush();
+    stacktrace_forked();
 
     if(nd_log.syslog.initialized) {
         closelog();
